SVM: Loader: add `validate_fee_payer` from `account_loader`
