+++ svm/src/account_loader.rs
@@ -1,7 +1,6 @@
 use {
     crate::{
         account_overrides::AccountOverrides,
-        account_rent_state::RentState,
         loader::Loader,
         nonce_info::{NonceFull, NoncePartial},
         transaction_error_metrics::TransactionErrorMetrics,
@@ -15,7 +14,6 @@ use {
         fee::FeeDetails,
         message::SanitizedMessage,
         native_loader,
-        nonce::State as NonceState,
         pubkey::Pubkey,
         rent::RentDue,
         rent_collector::{CollectedInfo, RentCollector, RENT_EXEMPT_RENT_EPOCH},
@@ -25,7 +23,6 @@ use {
         transaction::{Result, SanitizedTransaction, TransactionError},
         transaction_context::{IndexOfAccount, TransactionAccount},
     },
-    solana_system_program::{get_system_account_kind, SystemAccountKind},
     std::num::NonZeroUsize,
 };
 
@@ -96,60 +93,6 @@ pub fn collect_rent_from_account(
 
         CollectedInfo::default()
     }
-}
-
-/// Check whether the payer_account is capable of paying the fee. The
-/// side effect is to subtract the fee amount from the payer_account
-/// balance of lamports. If the payer_acount is not able to pay the
-/// fee, the error_metrics is incremented, and a specific error is
-/// returned.
-pub fn validate_fee_payer(
-    payer_address: &Pubkey,
-    payer_account: &mut AccountSharedData,
-    payer_index: IndexOfAccount,
-    error_metrics: &mut TransactionErrorMetrics,
-    rent_collector: &RentCollector,
-    fee: u64,
-) -> Result<()> {
-    if payer_account.lamports() == 0 {
-        error_metrics.account_not_found += 1;
-        return Err(TransactionError::AccountNotFound);
-    }
-    let system_account_kind = get_system_account_kind(payer_account).ok_or_else(|| {
-        error_metrics.invalid_account_for_fee += 1;
-        TransactionError::InvalidAccountForFee
-    })?;
-    let min_balance = match system_account_kind {
-        SystemAccountKind::System => 0,
-        SystemAccountKind::Nonce => {
-            // Should we ever allow a fees charge to zero a nonce account's
-            // balance. The state MUST be set to uninitialized in that case
-            rent_collector.rent.minimum_balance(NonceState::size())
-        }
-    };
-
-    payer_account
-        .lamports()
-        .checked_sub(min_balance)
-        .and_then(|v| v.checked_sub(fee))
-        .ok_or_else(|| {
-            error_metrics.insufficient_funds += 1;
-            TransactionError::InsufficientFundsForFee
-        })?;
-
-    let payer_pre_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
-    payer_account
-        .checked_sub_lamports(fee)
-        .map_err(|_| TransactionError::InsufficientFundsForFee)?;
-
-    let payer_post_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
-    RentState::check_rent_state_with_account(
-        &payer_pre_rent_state,
-        &payer_post_rent_state,
-        payer_address,
-        payer_account,
-        payer_index,
-    )
 }
 
 /// Collect information about accounts used in txs transactions and
@@ -442,13 +385,11 @@ mod tests {
             loader::Loader, nonce_info::NonceFull,
             transaction_account_state_info::TransactionAccountStateInfo,
         },
-        nonce::state::Versions as NonceVersions,
         solana_compute_budget::{compute_budget::ComputeBudget, compute_budget_processor},
         solana_program_runtime::loaded_programs::{ProgramCacheEntry, ProgramCacheForTxBatch},
         solana_sdk::{
             account::{Account, AccountSharedData, ReadableAccount, WritableAccount},
             bpf_loader_upgradeable,
-            epoch_schedule::EpochSchedule,
             feature_set::FeatureSet,
             hash::Hash,
             instruction::CompiledInstruction,
@@ -458,14 +399,13 @@ mod tests {
             },
             native_loader,
             native_token::sol_to_lamports,
-            nonce,
             pubkey::Pubkey,
             rent::Rent,
             rent_collector::{RentCollector, RENT_EXEMPT_RENT_EPOCH},
             rent_debits::RentDebits,
             reserved_account_keys::ReservedAccountKeys,
             signature::{Keypair, Signature, Signer},
-            system_program, system_transaction, sysvar,
+            system_transaction, sysvar,
             transaction::{Result, SanitizedTransaction, Transaction, TransactionError},
             transaction_context::{TransactionAccount, TransactionContext},
         },
@@ -952,149 +892,6 @@ mod tests {
         test(tx_not_set_limit, &result_default_limit);
         test(tx_set_limit_99, &result_requested_limit);
         test(tx_set_limit_0, &result_invalid_limit);
-    }
-
-    struct ValidateFeePayerTestParameter {
-        is_nonce: bool,
-        payer_init_balance: u64,
-        fee: u64,
-        expected_result: Result<()>,
-        payer_post_balance: u64,
-    }
-    fn validate_fee_payer_account(
-        test_parameter: ValidateFeePayerTestParameter,
-        rent_collector: &RentCollector,
-    ) {
-        let payer_account_keys = Keypair::new();
-        let mut account = if test_parameter.is_nonce {
-            AccountSharedData::new_data(
-                test_parameter.payer_init_balance,
-                &NonceVersions::new(NonceState::Initialized(nonce::state::Data::default())),
-                &system_program::id(),
-            )
-            .unwrap()
-        } else {
-            AccountSharedData::new(test_parameter.payer_init_balance, 0, &system_program::id())
-        };
-        let result = validate_fee_payer(
-            &payer_account_keys.pubkey(),
-            &mut account,
-            0,
-            &mut TransactionErrorMetrics::default(),
-            rent_collector,
-            test_parameter.fee,
-        );
-
-        assert_eq!(result, test_parameter.expected_result);
-        assert_eq!(account.lamports(), test_parameter.payer_post_balance);
-    }
-
-    #[test]
-    fn test_validate_fee_payer() {
-        let rent_collector = RentCollector::new(
-            0,
-            EpochSchedule::default(),
-            500_000.0,
-            Rent {
-                lamports_per_byte_year: 1,
-                ..Rent::default()
-            },
-        );
-        let min_balance = rent_collector.rent.minimum_balance(NonceState::size());
-        let fee = 5_000;
-
-        // If payer account has sufficient balance, expect successful fee deduction,
-        // regardless feature gate status, or if payer is nonce account.
-        {
-            for (is_nonce, min_balance) in [(true, min_balance), (false, 0)] {
-                validate_fee_payer_account(
-                    ValidateFeePayerTestParameter {
-                        is_nonce,
-                        payer_init_balance: min_balance + fee,
-                        fee,
-                        expected_result: Ok(()),
-                        payer_post_balance: min_balance,
-                    },
-                    &rent_collector,
-                );
-            }
-        }
-
-        // If payer account has no balance, expected AccountNotFound Error
-        // regardless feature gate status, or if payer is nonce account.
-        {
-            for is_nonce in [true, false] {
-                validate_fee_payer_account(
-                    ValidateFeePayerTestParameter {
-                        is_nonce,
-                        payer_init_balance: 0,
-                        fee,
-                        expected_result: Err(TransactionError::AccountNotFound),
-                        payer_post_balance: 0,
-                    },
-                    &rent_collector,
-                );
-            }
-        }
-
-        // If payer account has insufficient balance, expect InsufficientFundsForFee error
-        // regardless feature gate status, or if payer is nonce account.
-        {
-            for (is_nonce, min_balance) in [(true, min_balance), (false, 0)] {
-                validate_fee_payer_account(
-                    ValidateFeePayerTestParameter {
-                        is_nonce,
-                        payer_init_balance: min_balance + fee - 1,
-                        fee,
-                        expected_result: Err(TransactionError::InsufficientFundsForFee),
-                        payer_post_balance: min_balance + fee - 1,
-                    },
-                    &rent_collector,
-                );
-            }
-        }
-
-        // normal payer account has balance of u64::MAX, so does fee; since it does not  require
-        // min_balance, expect successful fee deduction, regardless of feature gate status
-        {
-            validate_fee_payer_account(
-                ValidateFeePayerTestParameter {
-                    is_nonce: false,
-                    payer_init_balance: u64::MAX,
-                    fee: u64::MAX,
-                    expected_result: Ok(()),
-                    payer_post_balance: 0,
-                },
-                &rent_collector,
-            );
-        }
-    }
-
-    #[test]
-    fn test_validate_nonce_fee_payer_with_checked_arithmetic() {
-        let rent_collector = RentCollector::new(
-            0,
-            EpochSchedule::default(),
-            500_000.0,
-            Rent {
-                lamports_per_byte_year: 1,
-                ..Rent::default()
-            },
-        );
-
-        // nonce payer account has balance of u64::MAX, so does fee; due to nonce account
-        // requires additional min_balance, expect InsufficientFundsForFee error if feature gate is
-        // enabled
-        validate_fee_payer_account(
-            ValidateFeePayerTestParameter {
-                is_nonce: true,
-                payer_init_balance: u64::MAX,
-                fee: u64::MAX,
-                expected_result: Err(TransactionError::InsufficientFundsForFee),
-                payer_post_balance: u64::MAX,
-            },
-            &rent_collector,
-        );
     }
 
     #[test]
